% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdata.R
\docType{data}
\name{expdata}
\alias{expdata}
\title{Arabidopsis expression atlas data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 54 rows and 100 columns.
}
\source{
\href{https://www.ebi.ac.uk/gxa/experiments/E-MTAB-7978/Downloads}{Expression Atlas}
}
\usage{
data(expdata)
}
\description{
Expression data from 54 tissues of A.thaliana, normalized as TPM values.
}
\examples{
data(expdata)
expdata[1:10,1:10] # view section of the small TPM matrix
}
\keyword{datasets}
